# Merge `new_data_frame()` and `new_bare_data_frame()`