--- a/NEWS.md
+++ b/NEWS.md
@@ -14,8 +14,6 @@
   #> 2
   ```
 
-* New `new_bare_data_frame()` is an extremely fast data frame constructor.
-
 * `vec_ptype()` now errors on scalar inputs (#807).
 
 * `vec_ptype_finalise()` is now recursive over all data frame types, ensuring