--- a/R/type-data-frame.R
+++ b/R/type-data-frame.R
@@ -20,13 +20,6 @@ new_data_frame <- function(x = list(), n = NULL, ..., class = NULL) {
 }
 new_data_frame <- fn_inline_formals(new_data_frame, "x")
 
-#' @rdname new_data_frame
-#' @export
-new_bare_data_frame <- function(x = list(), n = NULL) {
-  .Call(vctrs_new_bare_data_frame, x, n)
-}
-new_bare_data_frame <- fn_inline_formals(new_bare_data_frame, "x")
-
 
 # Light weight constructor used for tests - avoids having to repeatedly do
 # stringsAsFactors = FALSE etc. Should not be used in internal code as is