--- a/src/init.c
+++ b/src/init.c
@@ -87,7 +87,6 @@ extern SEXP vctrs_validate_name_repair_arg(SEXP);
 extern SEXP vctrs_validate_minimal_names(SEXP, SEXP);
 extern SEXP vctrs_as_names(SEXP, SEXP, SEXP);
 extern SEXP vctrs_is_partial(SEXP);
-extern SEXP vctrs_new_bare_data_frame(SEXP, SEXP);
 
 // Very experimental
 // Available in the API header
@@ -194,7 +193,6 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_validate_minimal_names",     (DL_FUNC) &vctrs_validate_minimal_names, 2},
   {"vctrs_as_names",                   (DL_FUNC) &vctrs_as_names, 3},
   {"vctrs_is_partial",                 (DL_FUNC) &vctrs_is_partial, 1},
-  {"vctrs_new_bare_data_frame",        (DL_FUNC) &vctrs_new_bare_data_frame, 2},
   {NULL, NULL, 0}
 };