As discussed on Slack, it might not be worth it to add an exported function.

`new_data_frame()` now uses `.External()` to build a pairlist of attributes. This makes it almost as fast for the bare case:

```r
bench::mark(
  new_data_frame(),
  new_bare_data_frame(),
  iterations = 1000000,
  check = FALSE
)[1:8]
#> # A tibble: 2 x 8
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`  n_itr  n_gc
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>  <int> <dbl>
#> 1 new_data_frame()         795ns    921ns   944023.        0B     22.7 999976    24
#> 2 new_bare_data_frame()    673ns    769ns  1204428.        0B     21.7 999982    18
```

And faster for the non-bare case:

```r
bench::mark(
  attrib = new_data_frame(foo = 1, bar = 2, baz = 3, quux = 4),
  iterations = 1000000,
  check = FALSE
)[1:8]

# Before

#> # A tibble: 1 x 8
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`  n_itr  n_gc
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>  <int> <dbl>
#> 1 attrib       1.66µs   2.09µs   438982.        0B     7.46 999983    17

# After

#> # A tibble: 1 x 8
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`  n_itr  n_gc
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>  <int> <dbl>
#> 1 attrib       1.25µs   1.54µs   567567.        0B     9.08 999984    16
```